feat: Generate SDK from blueprint only, removing OpenAPI parsing
Remove all TEMPORARY output-parity workarounds and their TODOs from the
codegen. The generator now consumes only the Blueprint from
@seamapi/blueprint; the raw OpenAPI spec is no longer parsed anywhere in
this repository's codegen (it is only passed to the blueprint plugin,
which builds the Blueprint from it).

- Delete codegen/lib/openapi/, map-python-type.ts, and endpoint-rules.ts.
- Filter endpoints on blueprint isUndocumented flags instead of raw
  x-undocumented/x-fern-sdk-group-name extensions and hardcoded path
  lists.
- Derive namespaces and parent/child class wiring from blueprint route
  and namespace paths, which now also wires deeply nested namespaces
  (seam.acs.encoders.simulate) to their parent class.
- Derive method parameters from endpoint.request.parameters and return
  types from endpoint.response via a new blueprint-format-to-Python-type
  mapping (codegen/lib/python-type.ts).
- Build models.py from blueprint resources, events, action attempts, and
  pagination, merging event and action attempt variant properties into
  single dataclasses.
- Fix the parameter sort comparator precedence so an explicit position
  sorts a parameter first.

Expected generated output changes:

- integer parameters and properties are now typed float (the blueprint
  collapses integer into number)
- new workspaces.update() method returning None
- seam.acs.encoders.simulate is now reachable
- {resource}_id parameter sorts first in get() methods
- PhoneRegistration model removed (blueprint omits it; it was referenced
  nowhere)
- ActionAttempt.result is now Dict[str, Any] with DeepAttrDict access
- Batch fields are now List[Dict[str, Any]]; Device.device_type is now
  str

+++ seam/routes/access_codes.py
@@ -231,9 +231,9 @@ class AccessCodes(AbstractAccessCodes):
         self,
         *,
         device_id: str,
-        max_code_length: Optional[int] = None,
-        min_code_length: Optional[int] = None,
-        supported_code_lengths: Optional[List[int]] = None
+        max_code_length: Optional[float] = None,
+        min_code_length: Optional[float] = None,
+        supported_code_lengths: Optional[List[float]] = None
     ) -> None:
         json_payload = {}
